Provide application-wide context and Persistent login state
This commit restructures the code a good amount. Now, in `App.tsx` we
create a context provider. This is necessary to store 'state
variables' application-wide. This is useful for global
state-management. For instance, when the user is logged in, their
state should be authenticated. With the context provider, every
component will get that update via `useContext`.

By providing context throughout the application and storing the JWT in
AsyncStorage, we can have a persistent login between page refreshes.